+++ turbo_sdk/signers/__init__.py
@@ -1,4 +1,5 @@
 from .ethereum import EthereumSigner
 from .arweave import ArweaveSigner
+from .solana import SolanaSigner
 
-__all__ = ["EthereumSigner", "ArweaveSigner"]
+__all__ = ["EthereumSigner", "ArweaveSigner", "SolanaSigner"]